// src/app/modules/branch/models/branch-view-response.model.ts
export class  BranchViewResponse {
        id: number;
        selected: string;
        company_ID: number;
        Branch_Type: string;
        Name: string;
        Country: string;
        City: string;
        Town: string;
        Street: string;
        Number: string;
        Telephone: number;
        Location: string;
        Status: number;
        Contact_Person: number;
        Manager: number;
        Capabilities: string;
        Building_No: number;
        Landline_Number: number;
        Fax_Number: number;
        Branch_Email_address: string;
        Branch_Website: string;
        Branch_Social_Media_IDs: number;
        Branch_Data_Access_Permissions: number;
        Branch_Data_Sharing_Permissions: number;
        created_at?: any;
        updated_at?: any;
    }